archived deals for sub realcash account.

+++ utils/utils.py
@@ -56,7 +56,7 @@ class Utils:
         ''' 以昨日收盘价计算涨停价格
         '''
         if zdf == 30:
-            return Decimal(str(lclose * 1.3)).quantize(Decimal('0.00'), ROUND_FLOOR)
+            return float(Decimal(str(lclose * 1.3)).quantize(Decimal('0.00'), ROUND_FLOOR))
         pdec = self.precious_decimal(precious)
         zprc = float(Decimal(str((int(round(lclose * 100, 0)) + lclose * zdf) / 100.0)).quantize(pdec, ROUND_HALF_UP))
         return zprc
@@ -66,7 +66,7 @@ class Utils:
         ''' 以昨日收盘价计算涨停价格
         '''
         if zdf == 30:
-            return Decimal(str(lclose * 0.7)).quantize(Decimal('0.00'), ROUND_CEILING)
+            return float(Decimal(str(lclose * 0.7)).quantize(Decimal('0.00'), ROUND_CEILING))
         pdec = self.precious_decimal(precious)
         dprc = float(Decimal(str((int(round(lclose * 100, 0)) - lclose * zdf) / 100.0)).quantize(pdec, ROUND_HALF_UP))
         return dprc
